Speedup cache initialization
Initalize cache using the ThreadPoolExecutor inherited from the tornado
ApiHandler (so we don't need to define a new one which could steal
resources from the main one).
Moreover, cache initialization is done during the login phase. In this
way, the cache will be ready right before the redirection to the app
view, without polling the server for its state.